worker/requirements.txt
aio_pika
psycopg2-binary
